requirements.txt
# Core dependencies
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
aiosqlite==0.19.0

# Image generation
Pillow==10.1.0
matplotlib==3.8.2

# PDF generation
reportlab==4.0.7

# Date handling
python-dateutil==2.8.2

# Utilities
aiofiles==23.2.1
